feat: add configurable GKI kernel identity hook
