--- .github/workflows/apply-kernel-identity.yml
@@ -1,59 +1,0 @@
-name: Apply kernel identity feature
-
-on:
-  push:
-    branches:
-      - feat/v2.5.7-final
-    paths:
-      - .github/workflows/apply-kernel-identity.yml
-
-permissions:
-  contents: write
-
-jobs:
-  apply:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Check out
-        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1
-        with:
-          ref: feat/v2.5.7-final
-          fetch-depth: 0
-
-      - name: Apply and verify source-level regressions
-        shell: bash
-        run: |
-          set -euo pipefail
-          python3 - <<'PY'
-          from pathlib import Path
-          path = Path('.agent/apply_kernel_identity.py')
-          text = path.read_text()
-          broken = 'if arguments.len() !in 4..=5 {'
-          fixed = 'if !(4..=5).contains(&arguments.len()) {'
-          if broken in text:
-              text = text.replace(broken, fixed, 1)
-          path.write_text(text)
-          PY
-          python3 .agent/apply_kernel_identity.py
-          python3 - <<'PY'
-          from pathlib import Path
-          path = Path('module/template/webroot/ux.js')
-          text = path.read_text()
-          marker = '    ];\n\n    for (const row of COMPLETE_CATALOG_ROWS) {'
-          row = '        ["Custom", "自定义", "Personalizado", "Benutzerdefiniert", "Пользовательский", "Kustom", "कस्टम", "مخصص"],\n'
-          if row not in text:
-              if marker not in text:
-                  raise SystemExit('complete catalog marker missing')
-              text = text.replace(marker, row + marker, 1)
-          path.write_text(text)
-          PY
-          node --check module/template/webroot/policy.js
-          node --check module/template/webroot/ux.js
-          for test in module/webui-tests/*.test.js; do node "$test"; done
-          cargo fmt --manifest-path rust/Cargo.toml --all
-          cargo fmt --manifest-path rust/Cargo.toml --all -- --check
-          git config user.name "Yiğit"
-          git config user.email "40565628+tryigit@users.noreply.github.com"
-          git add -A
-          git commit -m "feat: add configurable GKI kernel identity hook"
-          git push origin HEAD:feat/v2.5.7-final
