+++ module/src/main/cpp/kernel_identity.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace cleverestricky::kernel_identity {
+void configure(const char *payload);
+bool install_hooks_if_enabled();
+}
